Remove NBSP chars during normalization
Added NBSP (U+00A0) to the odd characters list.
This showed up in the IBM-pibs test data (not in the template).
Handling in normalization should be a good fix for here and future cases.
A couple of template prechecks were affected.
Also added a missing build flag and license to test files.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>